Added timeline controller functions

+++ src/main/java/timeLinear/models/timeEvent/TimeEventRepository.java
@@ -2,5 +2,5 @@ package timeLinear.models.timeEvent;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimeEventRepository extends JpaRepository<TimeEvent, Integer> {
+public interface TimeEventRepository extends JpaRepository<TimeEvent, Long> {
 }
